+++ app/routes/file-delete.js
@@ -1,6 +1,6 @@
 const { blobServiceClient } = require('../blob-storage')
 const storageConfig = require('../config/storage')
-const deleteMetadataFromCosmos = require('../message/delete-metadata-from-cosmos')
+const deleteMetadataFromCosmos = require('../storage/delete-metadata-from-cosmos')
 
 module.exports = {
   method: 'DELETE',
@@ -17,10 +17,13 @@ module.exports = {
       if (!exists) {
         return h.response({ error: 'File not found' }).code(404)
       }
-      await deleteMetadataFromCosmos(request, uniqueId)
+      const metadataResponse = await deleteMetadataFromCosmos(request, uniqueId)
 
       await blockBlobClient.delete()
-      return h.response({ message: 'File deleted successfully' }).code(200)
+      return h.response({
+        message: 'File deleted successfully',
+        metadata: metadataResponse
+      }).code(200)
     } catch (error) {
       console.error('Error deleting file:', error.message)
       return h.response({ error: 'File deletion failed', details: error.message }).code(500)
