+++ app/routes/file-upload.js
@@ -1,8 +1,5 @@
-const { checkFileExtension } = require('../utils/file-checks/extension-check')
 const { handleFileLimitExceeded } = require('../utils/file-checks/lenght-of-files-array-check')
-const { handleMessage } = require('../message')
-const { uploadFileToBlob } = require('../message/upload-file-to-blob')
-const { deleteBlob } = require('../message/delete-blob')
+const { handleFileUpload } = require('../storage/handle-file-upload')
 
 module.exports = {
   method: 'POST',
@@ -26,24 +23,8 @@ module.exports = {
       handleFileLimitExceeded(filesArray)
 
       for (const file of filesArray) {
-        let blockBlobClient
-        try {
-          checkFileExtension(file.hapi.filename)
-          const scanResult = { isSafe: true }
-          if (!scanResult.isSafe) {
-            results.push({ error: 'File is malicious and has been rejected', fileName: file.hapi.filename })
-            continue
-          }
-          const uploadResult = await uploadFileToBlob(file, payload)
-          blockBlobClient = uploadResult.blockBlobClient
-          await handleMessage({ body: uploadResult.metadata })
-          results.push({ message: 'File uploaded successfully', metadata: uploadResult.metadata })
-        } catch (error) {
-          if (blockBlobClient) {
-            await deleteBlob(blockBlobClient)
-          }
-          results.push({ error: `Failed to upload metadata, file: ${file.hapi.filename} is not saved` })
-        }
+        const result = await handleFileUpload(file, payload)
+        results.push(result)
       }
 
       return h.response(results).code(201)
