+++ app/storage/delete-blob.js
