+++ app/storage/delete-metadata-from-cosmos.js
